src/main/java/takanorig/example/maven/jdk7/DiamondOperator.java
package takanorig.example.maven.jdk7;

import java.util.ArrayList;
import java.util.LinkedHashMap;
import java.util.List;
import java.util.Map;

public class DiamondOperator
{
    public List<String> createList()
    {
        List<String> list = new ArrayList<>();
        list.add("test1");
        list.add("test2");
        list.add("test3");

        return list;
    }

    public Map<String, String> createMap()
    {
        Map<String, String> map = new LinkedHashMap<>();
        map.put("key1", "value1");
        map.put("key2", "value2");
        map.put("key3", "value3");

        return map;
    }
}
